+++ CloudNotes/CloudNotes/Extension/IntExtension.swift
@@ -8,18 +8,7 @@
 import Foundation
 
 extension Int {
-    struct Formatter {
-        static let utcFormatter: DateFormatter = {
-            let dateFormatter = DateFormatter()
-            dateFormatter.timeZone = TimeZone.autoupdatingCurrent
-            dateFormatter.locale = Locale.autoupdatingCurrent
-            dateFormatter.dateFormat = "yyyy.MM.dd"
-            
-            return dateFormatter
-        }()
-    }
-    
     var stringFromUTC: String? {
-        return Formatter.utcFormatter.string(from: Date(timeIntervalSince1970: Double(self)))
+        return CustomDateFormatter.utcFormatter.string(from: Date(timeIntervalSince1970: Double(self)))
     }
 }
